Modified the file robust_division_calculator.py

+++ programming_paradigm/robust_division_calculator.py
@@ -1,7 +1,7 @@
 def safe_divide(numerator, denominator):
     try:
-        numerator = int(numerator)
-        denominator = int(denominator)
+        numerator = float(numerator)
+        denominator = float(denominator)
         return f"The result of the division is {numerator / denominator}"
     except ZeroDivisionError:
         return "Error: Cannot divide by zero."
